python/requirements.txt
pandas
openpyxl
requests
python-dotenv
openai
pydantic
anthropic
